Refactor so static inner classes exist in their own files. Migrate some functionality to NumberUtilities.
