+++ open-sphere-base/analysis/src/main/java/io/opensphere/analysis/table/functions/SpreadsheetUtilities.java
@@ -6,9 +6,8 @@ package io.opensphere.analysis.table.functions;
 public final class SpreadsheetUtilities
 {
     /** Function values. */
-    private static final SpreadsheetFunction[] FUNCTION_VALUES = new SpreadsheetFunction[] { new SpreadsheetFunction.Sum(),
-        new SpreadsheetFunction.Min(), new SpreadsheetFunction.Max(), new SpreadsheetFunction.Median(),
-        new SpreadsheetFunction.Mean() };
+    private static final SpreadsheetFunction[] FUNCTION_VALUES = new SpreadsheetFunction[] { new Sum(), new Min(), new Max(),
+        new Median(), new Mean() };
 
     /**
      * Constructor.
